feature : 플레이어 공격 이펙트 추가 (임시)

+++ Project_Blind/Assets/Scripts/Core/StateMachineBehaviour/Player/MeleeAttackCombo3SMB.cs
@@ -30,7 +30,7 @@ namespace Blind
             {
                 if (_monoBehaviour.isJump)
                 {
-                    _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * _monoBehaviour.GetFacing());
+                    _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * (float)_monoBehaviour.GetFacing());
                 }
 
                 _monoBehaviour.enableAttack();
@@ -71,7 +71,7 @@ namespace Blind
                 animator.speed = 1.0f;
                 _monoBehaviour._attack.DamageReset(_monoBehaviour.Data.powerAttackdamage);
                 _monoBehaviour.enableAttack();
-                _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * _monoBehaviour.GetFacing());
+                _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * (float)_monoBehaviour.GetFacing());
                 _monoBehaviour.CurrentWaveGauge -= 10;
                 _monoBehaviour.isPowerAttackEnd = false;
                 if (ui == null)
